Move PlatformBindings to client.ts
